+++ packages/imperative/src/config/ConfigManagementFacility.ts
@@ -18,6 +18,7 @@ import { schemaDefinition } from "./cmd/schema/schema.definition";
 import { profilesDefinition } from "./cmd/profiles/profiles.definition";
 import { secureDefinition } from "./cmd/secure/secure.definition";
 import { setDefinition } from "./cmd/set/set.definition";
+import { editDefinition } from "./cmd/edit/edit.definition";
 import { importDefinition } from "./cmd/import/import.definition";
 import { convertProfilesDefinition } from "./cmd/convert-profiles/convert-profiles.definition";
 import { updateSchemasDefinition } from "./cmd/update-schemas/update-schemas.definition";
@@ -76,6 +77,7 @@ export class ConfigManagementFacility {
                 initDefinition,
                 schemaDefinition,
                 profilesDefinition,
+                editDefinition,
                 importDefinition,
                 convertProfilesDefinition,
                 updateSchemasDefinition