+++ packages/imperative/src/config/cmd/convert-profiles/convert-profiles.handler.ts
@@ -127,7 +127,6 @@ export default class ConvertProfilesHandler implements ICommandHandler {
         }
 
         const cliBin = ImperativeConfig.instance.rootCommandName;
-        // TODO Implement config edit command
         params.response.console.log(`Your new profiles have been saved to ${teamConfig.layerActive().path}.\n` +
             `Run "${cliBin} config edit --global-config" to open this file in your default editor.\n\n`);
 
@@ -206,6 +205,10 @@ export default class ConvertProfilesHandler implements ICommandHandler {
     /**
      * Disable the CredentialManager override in app settings. This is called
      * before uninstalling `ZOWE_CLI_SECURE_PLUGIN_NAME`.
+     * This method is private because only the convert-profiles command is able
+     * to disable the credential manager and reload it. For all other commands,
+     * the credential manager is loaded in `Imperative.init` and frozen with
+     * `Object.freeze` so cannot be modified later on.
      */
     private disableCredentialManager() {
         AppSettings.instance.set("overrides", "CredentialManager", ImperativeConfig.instance.hostPackageName);
@@ -215,7 +218,9 @@ export default class ConvertProfilesHandler implements ICommandHandler {
     }
 
     /**
-     * Uninstall a CLI plug-in if it is installed, otherwise do nothing.
+     * Uninstall a CLI plug-in if it is installed, otherwise do nothing. This
+     * overrides the default behavior of the plugin uninstall helper method
+     * which throws an error if the plugin is not installed.
      * @param pluginName Name of plug-in to uninstall
      */
     private uninstallPlugin(pluginName: string): void {